V1/requirements.txt
python-telegram-bot==13.7
Pillow==8.3.1